requirements.txt
numpy
pandas


matplotlib
seaborn
geopandas
scikit-learn

rpy2

tensorflow==2.20.0
tensorboard==2.20.0

fastapi
mysql-connector-python
uvicorn
python-multipart
#git lfs install